ActivityManagerService: Remove some duplication.
This was there because we want to keep both the ordering and the
consistency of the dump (in the non-client case). The fix is to
have a single function which drops the lock for the non-client
case; then in the client case, the caller acquires the lock,
ensuring that we still get a fully consistent dump in that case.

This change also makes the dumpClient case consistent with the
not-dumpClient case.

Test: builds, output looks sane
Change-Id: If54969e62789da02c0f2e3e19e75c036b189993b