Go via FgThread when starting the profiles during system boot.
Curently there is a race condition between unlocking the parent
user (spread across FgThread and mHandlerThread) and booting the
profile (happens on mHandlerThread). If the latter is too slow
and parent is still locked when the profile has reached
STATE_RUNNING_LOCKED, it won't proceed to unlock, which causes
the issue.

More complete RCA is on the first bug.

Bug:112830273
Bug:113128300
Test: atest UserControllerTest
Test: manual boot up with profile without passwords
Change-Id: I44dcbb431f97bd40ec00f417cd4b7e935a6b06bc